3.5.5 - waveform rewrite

+++ core/src/alchyr/taikoedit/editor/views/GameplayView.java
@@ -457,7 +457,7 @@ public class GameplayView extends MapView {
     }
 
     @Override
-    public float getBasePosition() {
+    public int getBasePosition() {
         return HIT_AREA_X;
     }
